Fix dependency cyle
